Reorganize core files into device and grid submodules
