+++ src/quatrex/core/device/contact.py
