+++ src/quatrex/core/grid/energies.py
