+++ src/quatrex/core/qtbm.py
@@ -28,8 +28,7 @@ from qttools.wave_function_solver import (
 from quatrex.core.compute_config import ComputeConfig
 from quatrex.core.constants import e, h
 from quatrex.core.device import Device
-from quatrex.core.energies import get_electron_energies
-from quatrex.core.kpoints import monkhorst_pack
+from quatrex.core.grid import get_electron_energies, monkhorst_pack
 from quatrex.core.quatrex_config import QuatrexConfig, SolverConfig
 from quatrex.core.statistics import fermi_dirac
